requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
line-bot-sdk==3.11.0
anthropic==0.28.0
python-telegram-bot==21.3
gspread==6.1.2
google-auth==2.29.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
